Factored non-DA parts of client into separate file
Eventually will probably move into it's own package, so people can just
`pip install` it, but for now they can manually copy things.
